refactor(via-router): prefer recursion + code cleanup
