+++ crates/via-router/Cargo.toml
@@ -9,7 +9,7 @@ homepage = "https://github.com/zacharygolba/via"
 repository = "https://github.com/zacharygolba/via"
 
 [features]
-default = []
+default = ["lru-cache"]
 lru-cache = []
 
 [dependencies]
